Moves method evaluation to the method factory
